Don't override php for composer in admin
I should have known this as I participated in some of the discussions.

Composer is now built with `buildComposerProject` which didn't exist in
the version of nixpkgs I'm pinning for php 8.1.9. Since I'm not in the
mood of doing more manual php patches, I'm rolling back overriding the
version of php in composer for admin.

Unfortunately this means that composer won't really work in admin for
now.